+++ src/components/creator/Editor/ChallengeDetailsEdition/ToolBoxEditor/toolboxPreview.css
